#5431 - Allow triggering bulk recommender via remote API
- Reorganize code to better separte original AERO from "next" version of the remote API while ensuring a smooth transition
