+++ inception/inception-remote/src/main/java/de/tudarmstadt/ukp/clarin/webanno/webapp/remoteapi/webhooks/json/DocumentStateChangeMessage.java
@@ -17,10 +17,11 @@
  */
 package de.tudarmstadt.ukp.clarin.webanno.webapp.remoteapi.webhooks.json;
 
+import static de.tudarmstadt.ukp.inception.remoteapi.SourceDocumentStateUtils.sourceDocumentStateToString;
+
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
-import de.tudarmstadt.ukp.clarin.webanno.webapp.remoteapi.aero.AeroProjectController;
 import de.tudarmstadt.ukp.inception.documents.event.DocumentStateChangedEvent;
 
 public class DocumentStateChangeMessage
@@ -51,9 +52,8 @@ public class DocumentStateChangeMessage
         documentId = aEvent.getDocument().getId();
         documentName = aEvent.getDocument().getName();
 
-        documentState = AeroProjectController.sourceDocumentStateToString(aEvent.getNewState());
-        documentPreviousState = AeroProjectController
-                .sourceDocumentStateToString(aEvent.getPreviousState());
+        documentState = sourceDocumentStateToString(aEvent.getNewState());
+        documentPreviousState = sourceDocumentStateToString(aEvent.getPreviousState());
     }
 
     public long getTimestamp()
